modified int removed bzeros

+++ ft_print_char.c
@@ -6,13 +6,13 @@
 /*   By: lgrobe-d <lgrobe-d@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2025/07/08 16:40:41 by lgrobe-d          #+#    #+#             */
-/*   Updated: 2025/07/14 14:34:37 by lgrobe-d         ###   ########.fr       */
+/*   Updated: 2025/07/16 12:14:57 by lgrobe-d         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "ft_printf.h"
 
-ssize_t	ft_print_char(va_list *ap)
+int	ft_print_char(va_list *ap)
 {
 	char	c;
